pyproject.toml
[project]
name = "finova-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "google-adk>=1.15.1",
    "ipykernel>=6.30.1",
    "langfuse>=3.6.1",
    "mcp>=1.16.0",
    "openinference-instrumentation-google-adk>=0.1.5",
]
